v4.1.0
* chore(package): Include GitHub repository

Included the GitHub repository for the project in the package.json file.

* fix(DM): Users with disabled Direct Messages

Fixed an issue where running a command that DM's a user who had Direct Messages disabled for server recipients would make the bot error.

* chore(version): Preparing for next semantic version

* chore(package): Update Dependencies

* minor(trivia): Lowering Cooldown

Due to popular demand, the trivia cooldown has now been lowered from 1 minute to 30 seconds.

* fix(interaction): DMing Slash Commands

Re-worked the fix for users DMing slash commands, resulting in errors. Prior to 4.0.1, the current solution was broken.

* fix(prompts): No Content

Inputting no message content or uploading videos/images for prompts that required message content resulted in errors. This commit fixes this for the report, bug-report and suggest command.

* chore(package): Updating to upcoming semantic version

Updating the package version to the next major release of v4.1.0

* feat(botstats): Command Rework

Reworked the bot stats command from the ground up. Includes:

- Refreshed Embed
- Database Connection Status
- Fixes to bot latency being API latency
- Including API latency to embed
- Bot version now updates according to package.json file
- Discord.js version is now displayed
- Acknowledgements code is more easily readable

* feat(messageReport): Report Message

A brand new context menu feature that allows users to easily report message content of other users to be reviewed by a staff member. Includes:

- Selection of messages to report (1, 3, 5 and 10)
- Visualise message content through HTML file
- Automatic posting to report abuse

* minor(botstats): Removing buttons on collector end

* fix(reportmessage): Catch Errors

Catching errors when removing the select menu.

* major(commands): Chat Commands Removal [WIP]

Removal of the chat command support for all of SaikouBot's commands, at the time of writing around 20-30 commands have been changed, next commit will include the rest.

This commit also includes full slash command support for some commands that didn't have it prior.

* feat(report): Command Rework

Completely revamped the report command, now includes:

- Full slash command support
- Use of select menus
- Use of forms
- Removal of the anonymous feature
- Embed refresh

* feat(suggest): Command Rework

Completely reworked the suggest command, now includes:

- Full slash command support
- Removal of the old clunky prompt system

* minor(anonymous): Removal of report

Removal of the report option for anonymous, all reports are now publicly visible due to message history permission changes in the server.

* feat(update): Command Rework

Completely reworked the update command, now includes:

- Rename to update, easier to type
- Fully slash command based
- Ability to accept/deny reports

* minor(embeds): Embed changes

* feat(commands): Full Slash Command Support

All of SaikouBot's commands now contain full slash command support! 🎉

* rework(bugreport): Command Rework

The bug report command has been reworked to use the same system as the reworked report command, offering better usability, ease of use and much more!

* fix(report): Handling timeout

Fixed not handling the timeout if a user didn't select an option for the platform, resulting in an error.

* feat(update): DM Suggestion Status

Like with reports, suggestions now DM the user the status of their suggestion and staff notes attached to them.

* chore(typings): Removed un-needed types

* feat(automod): Insult + Toxicity filter

Incorporated an insult and toxicity filter that now posts a message in the channel prompting users to report the offender if their content infringes on our rules.

Swear filter now no longer uses experimental either due to the poor detection it offered.

* delete(messageCreate): Merging into other file

Due to the chat commands being removed, the need for the messageCreate file in the cmd-handler is no longer needed.

Any code that is still relevant has now been moved to the discord-server messageCreate file instead.

* minor(ready): Changing bot status

Changing bot status to include the new prefix for the bot (/).

* minor(handler): Moving cooldown

Moving the cooldown code below the limited channel code, preventing users from being put on a cooldown if they ran a command in a wrong channel.

* minor(messageDelete): Specifying types

Specifying the types for the message created timestamp.

* minor(embeds): Removing Message Support

Removing the message support for all embeds, as well as removing embeds that are no longer used.

* removal(addtrivia): Removing Command

Removing the add trivia command, isn't really used at the moment and not worth the time re-making from scratch.

* cleanup(message): Remove logging

Removing console log that was accidentally committed during testing.

* fix(update): No Data Found

Fixed the update command not updating any suggestion or report due to not being able to find its data despite it existing.

* fix(trivia): Adding Emojis

Fixed a bug where users would add custom number emojis and it would count as an answer.

* fix(avatar): No Avatar

Fixed the avatar command erroring if a user with the default Discord avatar was used.

* minor(introduce): Form Uplift

Uplifted the introduce form to include placeholder text and minimum characters, which also assisted against people inputting empty spaces.

Also added a cancel embed for when the form times out.

* fix(suggest): Handling multiple prompts

Handling when a user tries to run the command multiple times before submitting the suggestion.

* fix(ban/warn): Ban Message History

Fixed the ban and warn command ban option message history not working due to it being renamed.

Also fixed if the option to delete message history wasn't inputted in time, correctly handled now.

* fix(punish): Timeout Errors

Fixed errors regarding the punish command timeout errors that would happen if a user didn't input an answer in time.

* fix(robloxban): Interaction Not Replied

Possible fix to the interaction not replied error that displays when one of the API's error.

* fix(reportMessage): Adding user data

Adding user data so reports done through the context menu can be approved/denied as well as removing the anonymous portion of the report

* feat(award): Developer award command

An easy way for developers to add points or booster tokens to a specific user.

* feat(emit): Emit Developer Command

A command used for emitting certain Discord events for testing.

* feat(redeem): Booster Redeem Command

The redeem command allows users who have acquired tokens through boosting to redeem them and informs a member of staff who they want to give the weapons to.

* fix(suggest): Interaction Acknowledged Error

* minor(types): Type Guards

Adding type guards for GuildMember instances as well as minor changes based on TypeScript intellisense.

* minor(embeds): Uplifts

Uplifting the embeds that either were just message content or poorly made.

* feat(suggestions): Deleting Suggestions

Deleting suggestions automatically that have 15+ downvotes, assisting in filtering the suggestions our community want and don't want.

* chore(packages): Update Dependencies

* feat(trivia): Weekly Trivia Leaderboard

Adding weekly trivia leaderboards for some extra competitiveness

* minor(trivia): Update non-existent user

* minor(reportMessage): Increase Cooldown

* minor(message): Delete Message Timeout

* fix(report): Invalid Roblox user message

* fix(trivia): Possible fixes

Possible fixes to the out of time error and interaction not being sent or deferred.